refactor(site/src): flatten chatCache into top-level cache operations
