+++ site/src/pages/AgentsPage/components/ChatConversation/useChatStore.ts
@@ -12,8 +12,9 @@ import {
 } from "react-query";
 import { watchChat } from "#/api/api";
 import {
-	chatCache,
 	chatMessagesKey,
+	invalidateChatPrompts,
+	patchChatMessages,
 	updateInfiniteChatsCache,
 } from "#/api/queries/chats";
 import type * as TypesGen from "#/api/typesGenerated";
@@ -40,7 +41,7 @@ const writeQueuedMessagesToCache = (
 		return;
 	}
 	const nextQueuedMessages = queuedMessages ?? [];
-	chatCache.patchMessages(queryClient, chatID, (currentData) => {
+	patchChatMessages(queryClient, chatID, (currentData) => {
 		if (!currentData?.pages?.length) {
 			return currentData;
 		}
@@ -195,7 +196,7 @@ export const useChatStore = (
 			if (!chatID || messages.length === 0) {
 				return;
 			}
-			chatCache.patchMessages(queryClient, chatID, (currentData) => {
+			patchChatMessages(queryClient, chatID, (currentData) => {
 				if (!currentData?.pages?.length) {
 					return currentData;
 				}
@@ -231,7 +232,7 @@ export const useChatStore = (
 			// Refresh the dedicated prompt-history cache when a user message arrives.
 			const hasNewUserPrompt = messages.some((msg) => msg.role === "user");
 			if (hasNewUserPrompt) {
-				void chatCache.invalidatePrompts(queryClient, chatID);
+				void invalidateChatPrompts(queryClient, chatID);
 			}
 		},
 		[chatID, queryClient],
@@ -242,7 +243,7 @@ export const useChatStore = (
 			if (!chatID) {
 				return;
 			}
-			chatCache.patchMessages(queryClient, chatID, (currentData) => {
+			patchChatMessages(queryClient, chatID, (currentData) => {
 				if (!currentData?.pages?.length) {
 					return currentData;
 				}
